3.8

+++ src/main/java/com/rzyou/funtime/common/jwt/constant/SecretConstant.java
@@ -11,8 +11,8 @@ public class SecretConstant {
     //签名秘钥
     public static final String BASE64SECRET = "ZW]4l5JH[m6Lm)LaQEjpb!4ElRaG(";
 
-    //超时毫秒数（默认30分钟）
-    public static final int EXPIRESSECOND = 30*24*60*60*1000;
+    //超时毫秒数（默认30天）
+    public static final long EXPIRESSECOND = 30*24*60*60*1000L;
 
     //用于JWT加密的密匙
     public static final String DATAKEY = "u^3y6SPER4jm*fn";
